Refactor Serializers with centralized abstract class
