+++ src/Serializer/SerializerInterface.php
@@ -7,11 +7,11 @@ namespace Sentry\Serializer;
 interface SerializerInterface
 {
     /**
-     * Serialize an object (recursively) into something safe to be sent in an Event
+     * Serialize an object (recursively) into something safe to be sent in an Event.
      *
      * @param mixed $value
-
-     * @return string|bool|float|int|null|array
+     *
+     * @return string|bool|float|int|null|object|array
      */
     public function serialize($value);
 }
